// src/app/shared/ged/enums/ged-enums.ts
export enum TFileType {
    EspelhoPonto = 1,
    Atestado = 2,
    VariaveisFolha = 3,
    Holerite = 4,
    Apontamento = 5,
    Obrigacao = 20,
    ArquivoGenerico = 100,
    Pasta = 255
}
export enum TFileSubType {
    ObrigacaoImposto = 20,
    ObrigacaoAcessoria = 21,
    ObrigacaoRelatorio = 22,
    ReciboSalario = 41,
    ReciboAdiantamento = 42,
    Recibo1Parcela13 = 43,
    Recibo2Parcela13 = 44,
    ReciboPlr = 45,
    ReciboProLabore = 46,
    ReciboAbono = 47
}

export enum TGedRefType {
    Nenhum = 0,
    Cadastro = 1,
    Funcionario = 2
}